Wed Jan 18 15:02:37 2012                        Michael Jennings (mej)
Convert the prop() wrapper method to use the parameter count rather
than references to distinguish between the presence or absence of a
value.  This removes the need for the "UNDEF" hack-around and makes
one-liner wrappers possible.

Add unit tests to illustrate possible techniques for using prop().
The goal is to promote code reuse and cut down on cut-and-pasted
functions and the errors which can result.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@705 3fa55bd7-8392-0410-b022-a3fc164e7bfc