Large monitor code re-work
Changed the monitor code around a lot...
Builds upon changes started in [807]

Removed::
  * monitor/src/globals.h

Added::
  * monitor/src/util.h
  * monitor/src/getstats.h
  * monitor/src/globals.h.in

The code has been broken into apart to not include other C source 
files now, so a seperate header has been created for util.c and
getstats.c.

The functions in util.c that require the, once global, sqlite3
database now take a DB pointer as the last argument. The database
is still global within wwmon_aggregator.c itself.

The file wwmon.db is no-longer created in the current directory,
but in the directory:

  `@LOCALSTATEDIR@/warewulf/`

`@LOCALSTATEDIR@` is replaced by the configure script when the
file globals.h is generated.

Building by hand would be like:

{{{
$ ./autogen.sh --prefix=/usr --sysconfdir=/etc --localstatedir=/var
}}}

The above would be the equivelant to building out an RPM with
the spec file and leaving all defaults.

Err... yeah...
-J



git-svn-id: https://warewulf.lbl.gov/svn/trunk@824 3fa55bd7-8392-0410-b022-a3fc164e7bfc