Update dynamic_hosts with "wwsh file sync dynamic_hosts"
Triggers an update to the WW-generated dynamic_hosts File
when File->sync is called for dynamic_hosts. This is done
by having Event/DynamicHosts.pm listen for a
"dynamic_hosts.update" event, which is thrown by
File->sync if the File name passed is "dynamic_hosts" and
the File has no origin.

This is useful when, for example, you have edited the
/etc/warewulf/hosts_template file but have no changes to
make to any Nodes.

git-svn-id: https://warewulf.lbl.gov/svn/trunk@1017 3fa55bd7-8392-0410-b022-a3fc164e7bfc